[Edit] 클래스 변경 → BaseView (#7)
